+++ src/driver/aurora-postgres/AuroraPostgresConnectionOptions.ts
@@ -1,9 +1,9 @@
-import { BaseConnectionOptions } from "../../connection/BaseConnectionOptions"
+import { BaseDataSourceOptions } from "../../data-source/BaseDataSourceOptions"
 
 /**
  * Postgres-specific connection options.
  */
-export interface AuroraPostgresConnectionOptions extends BaseConnectionOptions {
+export interface AuroraPostgresConnectionOptions extends BaseDataSourceOptions {
     /**
      * Database type.
      */