+++ src/driver/cordova/CordovaConnectionOptions.ts
@@ -1,9 +1,9 @@
-import { BaseConnectionOptions } from "../../connection/BaseConnectionOptions"
+import { BaseDataSourceOptions } from "../../data-source/BaseDataSourceOptions"
 
 /**
  * Sqlite-specific connection options.
  */
-export interface CordovaConnectionOptions extends BaseConnectionOptions {
+export interface CordovaConnectionOptions extends BaseDataSourceOptions {
     /**
      * Database type.
      */