+++ src/driver/mongodb/MongoConnectionOptions.ts
@@ -1,11 +1,11 @@
-import { BaseConnectionOptions } from "../../connection/BaseConnectionOptions"
+import { BaseDataSourceOptions } from "../../data-source/BaseDataSourceOptions"
 import { ReadPreference } from "./typings"
 
 /**
  * MongoDB specific connection options.
  * Synced with http://mongodb.github.io/node-mongodb-native/3.1/api/MongoClient.html
  */
-export interface MongoConnectionOptions extends BaseConnectionOptions {
+export interface MongoConnectionOptions extends BaseDataSourceOptions {
     /**
      * Database type.
      */
@@ -268,7 +268,7 @@ export interface MongoConnectionOptions extends BaseConnectionOptions {
      */
     readonly loggerLevel?: "error" | "warn" | "info" | "debug"
 
-    // Do not overwrite BaseConnectionOptions.logger
+    // Do not overwrite BaseDataSourceOptions.logger
     // readonly logger?: any;
 
     /**